Refactor parser and builtins into separate modules
