+++ index.html
@@ -150,6 +150,8 @@
     <script src="./src/mini4gl/statements/forEach.js"></script>
     <script src="./src/mini4gl/statements/find.js"></script>
     <script src="./src/mini4gl/statements/index.js"></script>
+    <script src="./src/mini4gl/parser.js"></script>
+    <script src="./src/mini4gl/builtins.js"></script>
     <script src="./mini4GL.js"></script>
     <script>
       const editor = document.getElementById('editor');
